feat(Exo13): panier avec store plus ou moins autonome

+++ client/src/components/DishesList/DishesList.tsx
@@ -1,10 +1,11 @@
 import { Box, Button, Stack, Typography } from "@mui/material";
+import { useCartStore, useToastStore } from "@/store";
+import { useEffect, useState } from "react";
 
 import CreateDishAccordion from "./CreateDishAccoridion";
 import DisplayDishAccordion from "./DisplayDishAccordion";
 import { dishApi } from "@/services/api";
 import useDishes from "@/hooks/useDishes";
-import { useState } from "react";
 
 interface DishesListProps {
   restaurantId: number;
@@ -24,26 +25,71 @@ const DishesList: React.FC<DishesListProps> = ({
   const handleCancelAddOrUpdate = () => {
     setCreating(false);
   };
+
+  const setToast = useToastStore((state) => state.setToast);
+  const { addToCart, increaseQuantity, decreaseQuantity, getQuantity } =
+    useCartStore((state) => state);
   const handleCreateDish = (toCreate: CreateDishDTO) => {
     dishApi.create(restaurantId, toCreate).then(
       () => {
         setCreating(false);
         refreshDishes();
+        setToast({
+          message: "Plat créé avec succès",
+          type: "success",
+          isOpen: true,
+        });
       },
       (error) => {
         console.error("Erreur lors de la création du plat:", error);
+        setToast({
+          message: "Erreur lors de la création du plat",
+          type: "error",
+          isOpen: true,
+        });
       }
     );
   };
   const handleDeleteDish = (dishId: number) => () => {
-    dishApi.delete(restaurantId, dishId).then(
-      () => {
-        refreshDishes();
-      },
-      (error) => {
-        console.error("Erreur lors de la suppression du plat:", error);
-      }
-    );
+    window.confirm("Voulez-vous vraiment supprimer ce plat ?") &&
+      dishApi.delete(restaurantId, dishId).then(
+        () => {
+          refreshDishes();
+          setToast({
+            message: "Plat supprimé avec succès",
+            type: "success",
+            isOpen: true,
+          });
+        },
+        (error) => {
+          console.error("Erreur lors de la suppression du plat:", error);
+          setToast({
+            message: "Erreur lors de la suppression du plat",
+            type: "error",
+            isOpen: true,
+          });
+        }
+      );
+  };
+  const handleAddToCart = (dish: DishDTO) => () => {
+    addToCart({
+      restaurantId: restaurantId,
+      dishId: dish.id,
+      name: dish.name,
+      price: dish.price,
+      quantity: 1,
+    });
+    setToast({
+      message: "Plat ajouté au panier",
+      type: "success",
+      isOpen: true,
+    });
+  };
+  const handleIncreaseQuantity = (dishId: number) => () => {
+    increaseQuantity(restaurantId, dishId);
+  };
+  const handleDecreaseQuantity = (dishId: number) => () => {
+    decreaseQuantity(restaurantId, dishId);
   };
 
   if (loading) return <Box>Chargement des plats...</Box>;
@@ -59,8 +105,13 @@ const DishesList: React.FC<DishesListProps> = ({
           <DisplayDishAccordion
             key={dish.id}
             dish={dish}
+            nbInCart={getQuantity(restaurantId, dish.id)}
             isEditMode={isEditMode}
-            handleDelete={handleDeleteDish(dish.id)}
+            handleClickAction={
+              isEditMode ? handleDeleteDish(dish.id) : handleAddToCart(dish)
+            }
+            handleIncreaseQuantity={handleIncreaseQuantity(dish.id)}
+            handleDecreaseQuantity={handleDecreaseQuantity(dish.id)}
           />
         ))}
         {isEditMode && creating && (
